Fix ESLint issues
Split organization context into separate files and update Tailwind config imports to resolve ESLint errors and enable Fast Refresh.
