+++ src/contexts/OrganizationContext.tsx
@@ -1,8 +1,9 @@
 
 // Compatibility layer - re-export everything from SimpleOrganizationContext
 export {
-  SimpleOrganizationProvider as OrganizationProvider,
-  type SimpleOrganizationContextType as OrganizationContextType
+  type SimpleOrganizationContextType as OrganizationContextType,
+  type SimpleOrganization
 } from './SimpleOrganizationContext';
 
+export { SimpleOrganizationProvider as OrganizationProvider } from './SimpleOrganizationProvider';
 export { useSimpleOrganization as useOrganization } from '@/hooks/useSimpleOrganization';
